\name{bst270Midus2024-package}
\alias{bst270Midus2024-package}
\alias{bst270Midus2024}
\docType{package}
\title{
\packageTitle{bst270Midus2024}
}
\description{
\packageDescription{bst270Midus2024}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bst270Midus2024}
\packageIndices{bst270Midus2024}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bst270Midus2024}

Maintainer: \packageMaintainer{bst270Midus2024}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
