% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_confounders.R
\name{filter_confounders}
\alias{filter_confounders}
\title{Clean and filter confounders}
\usage{
filter_confounders(df, group2_columns)
}
\arguments{
\item{df}{A dataframe.}

\item{group2_columns}{Confounder columns.}
}
\description{
Clean and filter confounders
}
\examples{
dat <- data.frame(sample_data)
filter_confounders(dat)
}
