% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_optimism.R
\name{filter_optimism}
\alias{filter_optimism}
\title{Clean and filter optimism measurements.}
\usage{
filter_optimism(df)
}
\arguments{
\item{df}{A dataframe.}
}
\description{
Clean and filter optimism measurements.
}
\examples{
df_raw <- data.frame(sample_survey_data)
df_optimism <- filter_optimism(df_raw)
}
