% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_table1_lipid.R
\name{gen_table1_lipid}
\alias{gen_table1_lipid}
\title{Display lipid measurement summary statistics by optimism level.}
\usage{
gen_table1_lipid(df)
}
\arguments{
\item{df}{A dataframe with combined survey and biomarker data.}
}
\description{
This function assumes pre-cleaned and filtered lipid measurements and optimism scores.
}
\examples{
df_raw <- data.frame(sample_survey_and_biomarker_data)
df_sum_stats_lipid <- gen_table1_lipid(df_raw)
}
