% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_table2.R
\name{gen_table2}
\alias{gen_table2}
\title{Display Table2: Associations between optimism and selected confounders}
\usage{
gen_table2(df)
}
\arguments{
\item{df}{A dataframe with surveydata.}
}
\description{
This function assumes pre-cleaned and filtered confounder and optimism scores.
}
\examples{
df_raw <- data.frame(sample_surveydata)
gen_table2(df_raw)
}
