% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pathway.R
\name{get_alchol_consumption}
\alias{get_alchol_consumption}
\title{Get the alcohol consumption}
\usage{
get_alchol_consumption(dat)
}
\arguments{
\item{dat}{A dataframe}
}
\value{
data including the alcohol consumption as average number of drinks per week
}
\description{
Get the alcohol consumption
}
